Implement minor enhancement features
- Delay for autoclosing BSOD after completion
- Delay for progress counter start
- Factor for slowing down/speeding up the progress bar
- Remove enable-unsafe flag as it was broken and inhibited using the
  command feature

+++ BluescreenSimulator/Models/Windows10Bluescreen.cs
@@ -21,7 +21,7 @@ namespace BluescreenSimulator
 
         public bool UseOriginalQR { get; set; } = true;
 		
-        public float TextDelay { get; set; }
+        public double TextDelay { get; set; }
 
         public bool CustomQR { get; set; } = false;
